+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v2.0.0-beta.38 - 2025-02-07
+
+- Follow Nexus changes
+
+## v2.0.0-beta.37 - 2025-02-07
+
+- Follow Nexus changes
+
 ## v2.0.0-beta.36 - 2025-01-22
 
 - Follow Nexus changes